requirements.txt
pandas
numpy
scikit-learn
wandb
fastapi
uvicorn
streamlit
pydantic
pyyaml
requests
dvc
